+++ src/main/java/com/gmorgan/framedesigner/repositories/SectionRepository.java
@@ -11,7 +11,7 @@ public interface SectionRepository extends JpaRepository<Section, Long> {
 
     List<Section> findAllByDescContainsOrderByIxx (String string);
     List<Section> findAllByIyyGreaterThanAndZyyGreaterThanAndDescContainsIgnoreCase(double iyy, double zyy, String desc);
-    List<Section> findAllByIxxGreaterThanAndZxxGreaterThanAndIyyGreaterThanAndZyyGreaterThanAndYGreaterThanAndDescContainsIgnoreCase(double ixx, double zxx, double iyy, double zyy, double x, String desc);
+    List<Section> findAllByIxxGreaterThanAndZxxGreaterThanAndIyyGreaterThanAndZyyGreaterThanAndYGreaterThanAndDescContainsIgnoreCase(double ixx, double zxx, double iyy, double zyy, int x, String desc);
     List<Section> findAllByIyyGreaterThanAndZyyGreaterThanAndCsaGreaterThanAndDescContainsIgnoreCase(double iyy, double zyy, double csa, String desc);
     List<Section> findAllByIyyGreaterThanAndZyyGreaterThanAndXGreaterThanAndYGreaterThanAndDescContainsIgnoreCase(double iyy, double zyy, int x, int y, String desc);
 }
